Restructure run scripts to avoid cyclical dependencies
And add refinements to naming, etc.
