+++ _tools/run/helpers/helpers.js
@@ -445,7 +445,6 @@ async function renderIndexComments (argv) {
   'use strict'
 
   if (projectSettings()['dynamic-indexing'] !== false) {
-
     console.log('Processing indexing comments ...')
 
     try {
@@ -476,7 +475,6 @@ async function renderIndexLinks (argv) {
   'use strict'
 
   if (projectSettings()['dynamic-indexing'] !== false) {
-
     console.log('Adding links to reference indexes ...')
 
     try {
@@ -830,14 +828,23 @@ async function runPrince (argv) {
       console.log('Using PrinceXML licence found at ' + princeLicenseFile)
     }
 
+    // Check if we're using the correct Prince version
     checkPrinceVersion()
+
+    // Get the HTML file to render. If we are merging
+    // input files, we only pass the merged file to Prince.
+    // Unless `--merged false` was passed at the command line.
+    let inputFiles = [fsPath.dirname(htmlFilePaths(argv)[0]) + '/merged.html']
+    if (argv.merged === false) {
+      inputFiles = htmlFilePaths(argv)
+    }
 
     // Currently, node-prince does not seem to
     // log its progress to stdout. Possible WIP:
     // https://github.com/rse/node-prince/pull/7
     prince()
       .license('./' + princeLicenseFile)
-      .inputs(htmlFilePaths(argv))
+      .inputs(inputFiles)
       .output(process.cwd() + '/_output/' + outputFilename(argv))
       .option('javascript')
       .option('verbose')
@@ -1250,7 +1257,6 @@ async function convertHTMLtoWord (argv) {
   // Clear the previous output folder if it exists,
   // or create the output directory first if it doesn't.
   if (pathExists(outputLocation)) {
-    // await clearFolderContents(outputLocation)
     await fs.emptyDir(outputLocation)
   } else {
     await fs.mkdir(outputLocation, { recursive: true })
@@ -1296,130 +1302,11 @@ async function convertHTMLtoWord (argv) {
   })
 }
 
-// Web output
-async function web (argv) {
-  'use strict'
-
-  try {
-    // await clearFolderContents(process.cwd() + '/_site')
-    await fs.emptyDir(process.cwd() + '/_site')
-    await jekyll(argv)
-  } catch (error) {
-    console.log(error)
-  }
-}
-
-// PDF output
-async function pdf (argv) {
-  'use strict'
-
-  try {
-    // await clearFolderContents(process.cwd() + '/_site')
-    await fs.emptyDir(process.cwd() + '/_site')
-    await jekyll(argv)
-    await renderMathjax(argv)
-    await renderIndexComments(argv)
-    await renderIndexLinks(argv)
-    await runPrince(argv)
-    openOutputFile(argv)
-  } catch (error) {
-    console.log(error)
-  }
-}
-
-// Epub output
-async function epub (argv) {
-  'use strict'
-
-  try {
-    // await clearFolderContents(process.cwd() + '/_site')
-    await fs.emptyDir(process.cwd() + '/_site')
-    await jekyll(argv)
-    await renderIndexComments(argv)
-    await renderIndexLinks(argv)
-    await convertXHTMLLinks(argv)
-    await convertXHTMLFiles(argv)
-    await cleanHTMLFiles(argv)
-    await addToEpub(htmlFilePaths(argv, '.xhtml'), argv.book)
-    await addToEpub(bookAssetPaths(argv, 'images'),
-      argv.book + '/images/epub')
-    await addToEpub(bookAssetPaths(argv, 'styles'),
-      argv.book + '/styles')
-    await addToEpub(bookAssetPaths(argv, 'images', 'assets'),
-      'assets/images/epub')
-
-    if (pathExists(process.cwd() + '/_site/assets/js/bundle.js')) {
-      await addToEpub([process.cwd() + '/_site/assets/js/bundle.js'],
-        '/assets/js')
-    }
-
-    if (mathjaxEnabled(argv)) {
-      await addToEpub([process.cwd() + '/_site/assets/js/mathjax'],
-        '/assets/js/mathjax')
-    }
-
-    await addToEpub([process.cwd() + '/_site/' +
-                argv.book + '/package.opf'], '')
-
-    const ncxFile = process.cwd() + '/_site/' +
-                argv.book + '/toc.ncx'
-    if (pathExists(ncxFile)) {
-      await addToEpub([ncxFile], '')
-    }
-    await epubZip()
-    await epubZipRename(argv)
-
-    const pathToEpub = fsPath.normalize(process.cwd() +
-      '/_output/' +
-      argv.book + '.epub')
-    await epubValidate(pathToEpub)
-
-    console.log('Your epub is at ' + pathToEpub)
-  } catch (error) {
-    console.log(error)
-  }
-}
-
-// App output
-async function app (argv) {
-  'use strict'
-
-  try {
-    // await clearFolderContents(process.cwd() + '/_site')
-    await fs.emptyDir(process.cwd() + '/_site')
-    await jekyll(argv)
-    await fsPromises.mkdir(process.cwd() + '/_site/app/www')
-    await assembleApp()
-
-    if (argv['app-build']) {
-      await cordova(['platform', 'add', argv['app-os']])
-      await cordova(['platform', 'prepare', argv['app-os']])
-
-      // Build the app
-      if (argv['app-release']) {
-        await cordova(['build', argv['app-os']], '--release')
-      } else {
-        await cordova(['build', argv['app-os']])
-      }
-
-      // Run emulator
-      if (argv['app-emulate']) {
-        await cordova(['emulate', argv['app-os']])
-      }
-    } else {
-      console.log('App folders ready in _site/app.')
-    }
-  } catch (error) {
-    console.log(error)
-  }
-}
-
 // Word export
 async function exportWord (argv) {
   'use strict'
 
   try {
-    // await clearFolderContents(process.cwd() + '/_site')
     await fs.emptyDir(process.cwd() + '/_site')
     await jekyll(argv)
 
@@ -1440,7 +1327,6 @@ async function refreshIndexes (argv) {
   'use strict'
 
   try {
-    // await clearFolderContents(process.cwd() + '/_site')
     await fs.emptyDir(process.cwd() + '/_site')
     await jekyll(argv)
 
@@ -1499,18 +1385,32 @@ function newBook (argv) {
 }
 
 module.exports = {
-  app,
-  pdf,
-  web,
-  epub,
+  addToEpub,
+  assembleApp,
+  bookAssetPaths,
+  bookIsValid,
+  cleanHTMLFiles,
+  convertXHTMLFiles,
+  convertXHTMLLinks,
+  cordova,
+  epubValidate,
+  epubZip,
+  epubZipRename,
   exportWord,
-  processImages,
+  htmlFilePaths,
   installGems,
   installNodeModules,
+  jekyll,
   logProcess,
+  mathjaxEnabled,
   newBook,
+  openOutputFile,
   pathExists,
+  processImages,
   refreshIndexes,
-  bookIsValid,
+  renderIndexComments,
+  renderIndexLinks,
+  renderMathjax,
+  runPrince,
   works
 }
